Fix hundreds of compile warnings.

+++ poly2tri/sweep/sweep_context.h
@@ -157,7 +157,7 @@ inline AdvancingFront* SweepContext::front()
 
 inline int SweepContext::point_count()
 {
-  return points_.size();
+  return (int)points_.size();
 }
 
 inline void SweepContext::set_head(Point* p1)
